+++ src/game/tests.rs
@@ -448,6 +448,7 @@ mod vintage_cube_plagon;
 mod vintage_cube_ponder;
 mod vintage_cube_portable_hole;
 mod vintage_cube_portal_to_phyrexia;
+mod vintage_cube_preacher_of_the_schism;
 mod vintage_cube_prismatic_ending;
 mod vintage_cube_psychic_frog;
 mod vintage_cube_quantum_riddler;